requirements.txt
braintree==3.53.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.7
django-allauth==0.36.0
django-crispy-forms==1.8.0
gunicorn==20.0.4
idna==2.7
oauthlib==2.1.0
Pillow==6.2.1
python3-openid==3.1.0
pytz==2018.4
requests==2.21.0
requests-oauthlib==1.0.0
sqlparse==0.3.0
stripe==1.82.1
urllib3==1.24.2
